chore: Update npm dependencies and add PostCSS plugins
The package.json file was modified to include the latest versions of the "postcss-import" and "postcss-nesting" dependencies. This update will improve the functionality of the application.

Additionally, the tailwind.config.js file was updated to extend the theme with new colors for the application.

The Login page component was also modified to include a Navbar component and a form with input fields for email and password, as well as a login button and a link to create a new account.

These changes were made to enhance the functionality and user experience of the application.